Removes the platform dependence in tests
The FileField and ImageField tests currently assume the existence of
a UNIX-style temp directory and the use of UNIX-style paths. On
Windows this causes the tests to fail and to create spurious new
files and directories.

This commit makes the tests platform-independent by using the
standard Python modules tempfile and os.path.