Recipe.prepare does not save related recipes
You can call `Recipe.prepare` on a model with a ForeignKey and if
you don't have defined a `foreign_key`-recipe the related object
will not be persited. However if you define a related recipe the
related object will be persistet.

Besides that behavior being inconsistent, there really is no reason
and it presents disadvantages.

This patch solves the issue, by using the `_save_realted` attribute
that is already existing but not used for related objects in recipes.