Avoid touching the database if not necessary
If I'm not mistaken, the intention of this bit of code is to store a backup copy of the given iterator while being able to iterate through it for this `make()` or `prepare()` call. We need to detect if this is the first run, so we check if any instances are present or if the dictionary is empty.

This causes a problem when using `prepare()`, because it still does a call to the database - something we usually don't want when using `prepare()`. If I'm not mistaken, we can safely switch the statement around, which avoids the call to the database.

To be honest, my gut feeling is that the call to `m.objects.count()` can be avoided all together - the existing logic should be sufficient. But I'm playing it safe here ;)