Add _from_manager option to return instance from custom Manager in make()
Prior behaviour is unchanged with _from_manager=None by default. This
feature helps to ensure that the instance returned by mommy.make() has the
same fields as a 'normal' instance, which may e.g. use a custom model
Manager bound to objects on the model, that may annotate additional fields.